requirements.txt
flask==3.0.2
networkx==3.2.1
pyvis==0.3.2
pyyaml==6.0.1